--- a/aten/src/ATen/native/Linear.cpp
+++ b/aten/src/ATen/native/Linear.cpp
@@ -1,7 +1,9 @@
 #include <ATen/ATen.h>
 #include <ATen/native/Resize.h>
+#include <ATen/native/mkldnn/Utils.h>
 #include <ATen/NativeFunctions.h>
 #include <ATen/native/xnnpack/Engine.h>
+#include <ATen/Config.h>
 #include <ATen/WrapDimUtilsMulti.h>
 #include <c10/macros/Macros.h>
 #include <c10/util/irange.h>
@@ -22,9 +24,11 @@ Tensor linear(const Tensor& input, const Tensor& weight, const c10::optional<Ten
     ? c10::MaybeOwned<Tensor>::borrowed(*bias_opt)
     : c10::MaybeOwned<Tensor>::owned(c10::in_place);
 
-  if (input.is_mkldnn()) {
+#if AT_MKLDNN_ENABLED()
+  if (input.is_mkldnn() || (input.scalar_type() == kBFloat16 && mkldnn_bf16_device_check())) {
     return at::mkldnn_linear(input, weight, *bias);
   }
+#endif
 #if defined(C10_MOBILE)
   if (xnnpack::use_linear(input, weight, *bias)) {
     return xnnpack::linear(input, weight, *bias);