--- a/aten/src/ATen/native/mkldnn/Linear.cpp
+++ b/aten/src/ATen/native/mkldnn/Linear.cpp
@@ -27,7 +27,9 @@ std::tuple<Tensor, Tensor, Tensor> mkldnn_linear_backward(
     const Tensor& weight, std::array<bool,3> output_mask) {
   TORCH_CHECK(false, "mkldnn_linear_backward: ATen not compiled with MKLDNN support");
 }
-
+Tensor mkldnn_linear_weight_pack(const Tensor& weight) {
+  TORCH_CHECK(false, "mkldnn_linear_weight_pack: ATen not compiled with MKLDNN support");
+}
 } // namespace native
 } // namespace at
 
@@ -41,7 +43,8 @@ namespace native {
 
 Tensor mkldnn_linear(
     const Tensor& self,
-    const Tensor& weight_t, const c10::optional<Tensor>& bias_opt) {
+    const Tensor& weight_t,
+    const c10::optional<Tensor>& bias_opt) {
   // See [Note: hacky wrapper removal for optional tensor]
   c10::MaybeOwned<Tensor> bias_maybe_owned = at::borrow_from_optional_tensor(bias_opt);
   const Tensor& bias = *bias_maybe_owned;
@@ -51,8 +54,6 @@ Tensor mkldnn_linear(
       self.dim() != 0,
       "mkldnn_linear: input needs to has dim at least 1, input dim ",
       self.dim());
-  TORCH_CHECK(self.is_mkldnn(),
-      "mkldnn_linear: input needs to be mkldnn layout");
   if (self.scalar_type() == ScalarType::BFloat16) {
     TORCH_CHECK(mkldnn_bf16_device_check(),
         "mkldnn_linear: bf16 path needs the cpu support avx512bw, avx512vl and avx512dq");
@@ -62,12 +63,18 @@ Tensor mkldnn_linear(
   auto self_reshaped =
       dim == 2 ? self : self.reshape({-1, self.size(self.dim() - 1)});
 
-  const ideep::tensor x = itensor_from_mkldnn(self_reshaped);
+  const ideep::tensor& x = itensor_from_tensor(self_reshaped);
   // weight_t can be a mkldnn tensor or dense tensor.
   const Tensor weight = (weight_t.is_mkldnn() || weight_t.is_contiguous()) ? weight_t : weight_t.contiguous();
-  const ideep::tensor w = itensor_from_tensor(weight);
+  const ideep::tensor& w = itensor_from_tensor(weight);
 
   ideep::tensor y;
+  Tensor aten_result;
+  if (!self.is_mkldnn()){
+    // aten in, aten out, share aten buffer to avoid copy
+    aten_result = at::empty({self_reshaped.size(0), weight.size(0)}, self.options());
+    y = itensor_from_tensor(aten_result);
+  }
   if (bias.defined()) {
     const ideep::tensor b = itensor_from_tensor(bias);
     ideep::inner_product_forward::compute(x, w, b, y);
@@ -79,25 +86,34 @@ Tensor mkldnn_linear(
   std::vector<int64_t> output_size(input_size.begin(), input_size.end() - 1);
   output_size.push_back(weight.size(0));
 
-  if (self.dim() != 2) {
-    return new_with_itensor_mkldnn(std::move(y), optTypeMetaToScalarType(self.options().dtype_opt()),
-                                   self.options().device_opt()).reshape(output_size);
+  if (self.is_mkldnn()){
+    // mkldnn in, mkldnn out
+    aten_result = new_with_itensor_mkldnn(std::move(y), optTypeMetaToScalarType(self.options().dtype_opt()),
+                                    self.options().device_opt());
   }
-  return new_with_itensor_mkldnn(std::move(y), optTypeMetaToScalarType(self.options().dtype_opt()),
-                                 self.options().device_opt());
+
+  if (self.dim() != 2) {
+    aten_result = aten_result.reshape(output_size);
+  }
+  return aten_result;
 }
 
 
 Tensor mkldnn_linear_backward_input(
     IntArrayRef input_size, const Tensor& grad_output, const Tensor& weight_t){
-  TORCH_CHECK(grad_output.is_mkldnn(),
-      "mkldnn_linear_backward: grad_output needs to be mkldnn layout");
-  TORCH_CHECK(weight_t.device().is_cpu() && weight_t.scalar_type() == kFloat,
-      "mkldnn_linear_backward: weight_t needs to be a dense tensor");
+  // TORCH_CHECK(grad_output.is_mkldnn(),
+  //     "mkldnn_linear_backward: grad_output needs to be mkldnn layout");
+  // TORCH_CHECK(weight_t.device().is_cpu() && weight_t.scalar_type() == kFloat,
+  //     "mkldnn_linear_backward: weight_t needs to be a dense tensor");
+  if (weight_t.scalar_type() == ScalarType::BFloat16) {
+    TORCH_CHECK(mkldnn_bf16_device_check(),
+        "mkldnn_linear: bf16 path needs the cpu support avx512bw, avx512vl and avx512dq");
+  }
+
   auto grad_output_reshaped = grad_output.dim() > 2 ?
     grad_output.reshape({-1, grad_output.size(grad_output.dim() - 1)}) : grad_output;
 
-  ideep::tensor& grady = itensor_from_mkldnn(grad_output_reshaped);
+  const ideep::tensor& grady = itensor_from_tensor(grad_output_reshaped);
   // weight_t always dense tensor for training.
   const Tensor weight = weight_t.is_contiguous() ? weight_t : weight_t.contiguous();
   const ideep::tensor w = itensor_view_from_dense(weight);
@@ -106,45 +122,44 @@ Tensor mkldnn_linear_backward_input(
   input_reshaped_size.push_back(grad_output_reshaped.size(0));
   input_reshaped_size.push_back(weight.size(1));
 
-  ideep::tensor gradx;
+  Tensor aten_gradx = at::empty(input_reshaped_size, grad_output.options());
+  ideep::tensor gradx = itensor_from_tensor(aten_gradx);
   ideep::inner_product_backward_data::compute(
     grady, w, {input_reshaped_size.begin(), input_reshaped_size.end()}, gradx);
 
-  if (input_size.size() > 2) {
-    return new_with_itensor_mkldnn(std::move(gradx), optTypeMetaToScalarType(grad_output.options().dtype_opt()),
-                                   grad_output.options().device_opt()).reshape(input_size);
+  if (grad_output.is_mkldnn()){
+    // mkldnn in, mkldnn out
+    aten_gradx = new_with_itensor_mkldnn(std::move(gradx), optTypeMetaToScalarType(grad_output.options().dtype_opt()),
+                                    grad_output.options().device_opt());
   }
-  return new_with_itensor_mkldnn(std::move(gradx), optTypeMetaToScalarType(grad_output.options().dtype_opt()),
-                                 grad_output.options().device_opt());
+
+  if (input_size.size() != 2) {
+    aten_gradx = aten_gradx.reshape(input_size);
+  }
+
+  return aten_gradx;
 }
 
 std::tuple<Tensor, Tensor> mkldnn_linear_backward_weights(
     const Tensor& grad_output, const Tensor& input, const Tensor& weight, bool bias_defined) {
-  TORCH_CHECK(grad_output.is_mkldnn() && input.is_mkldnn(),
-      "mkldnn_linear_backward: grad_output and input needs to be mkldnn layout");
-  TORCH_CHECK(weight.device().is_cpu() && weight.scalar_type() == kFloat,
-      "mkldnn_linear_backward: weight needs to be a dense tensor");
-
   auto grad_output_reshaped = grad_output.dim() > 2 ?
     grad_output.reshape({-1, grad_output.size(grad_output.dim() - 1)}) : grad_output;
   auto input_reshaped = input.dim() > 2 ? input.reshape({-1, input.size(input.dim() - 1)}) : input;
 
-  ideep::tensor& grady = itensor_from_mkldnn(grad_output_reshaped);
-  ideep::tensor& x = itensor_from_mkldnn(input_reshaped);
-  ideep::tensor gradw, gradb;
+  const ideep::tensor& grady = itensor_from_tensor(grad_output_reshaped);
+  const ideep::tensor& x = itensor_from_tensor(input_reshaped);
+
+  Tensor aten_gradw = at::empty(weight.sizes(), weight.options());
+  Tensor aten_gradb = bias_defined ? at::empty({weight.size(0)}, weight.options()) : Tensor();
+  ideep::tensor gradw = itensor_from_tensor(aten_gradw);
+  ideep::tensor gradb = itensor_from_tensor(aten_gradb);
   if (bias_defined) {
     ideep::inner_product_backward_weights::compute(x, grady, gradw, gradb);
   } else {
     ideep::inner_product_backward_weights::compute(x, grady, gradw);
   }
-
-  return std::tuple<Tensor, Tensor>{
-    mkldnn_to_dense(new_with_itensor_mkldnn(std::move(gradw),
-                    optTypeMetaToScalarType(weight.options().dtype_opt()),
-                    weight.options().device_opt())),
-    mkldnn_to_dense(new_with_itensor_mkldnn(std::move(gradb),
-                    optTypeMetaToScalarType(weight.options().dtype_opt()),
-                    weight.options().device_opt()))};
+  // grad_weight and grad_bias always dense tensor for training.
+  return std::tuple<Tensor, Tensor>{aten_gradw, aten_gradb};
 }
 
 std::tuple<Tensor, Tensor, Tensor> mkldnn_linear_backward(
@@ -158,6 +173,24 @@ std::tuple<Tensor, Tensor, Tensor> mkldnn_linear_backward(
     std::tie(grad_weight, grad_bias) = at::mkldnn_linear_backward_weights(grad_output, input, weight, output_mask[2]);
   }
   return std::tuple<Tensor, Tensor, Tensor>{grad_input, grad_weight, grad_bias};
+}
+
+Tensor mkldnn_linear_weight_pack(const Tensor& weight) {
+  TORCH_CHECK(weight.is_mkldnn(), "mkldnn_linear_weight_pack only pack mkldnn tensor");
+  const ideep::tensor w = itensor_from_tensor(weight);
+  auto expected_w_desc = ideep::inner_product_forward::expected_weights_desc(
+    w.get_dims(),
+    /*dummy src dims*/ideep::dims(),
+    w.get_data_type(),
+    /*assume src dtype should same with weight*/w.get_data_type()
+  );
+  if (w.get_desc() == expected_w_desc){
+    return weight;
+  }
+  ideep::tensor expected_w {expected_w_desc};
+  expected_w.feed_from(w);
+  return new_with_itensor_mkldnn(std::move(expected_w), optTypeMetaToScalarType(weight.options().dtype_opt()),
+                                    weight.options().device_opt());
 }
 
 } // namespace native