… aten/src/ATen/native/native_functions.yaml
… test/test_mkldnn.py
… torch/_C/_nn.pyi.in